Move cli setup functions to controller
