+++ src/smif/cli/__init__.py
@@ -74,7 +74,6 @@ import logging
 import logging.config
 import os
 import pkg_resources
-import shutil
 
 try:
     import win32api
@@ -88,6 +87,7 @@ import smif
 import smif.cli.log
 
 from smif.controller.execute import execute_model_run
+from smif.controller.setup import setup_project_folder
 from smif.http_api import create_app
 from smif.data_layer import DatafileInterface
 
@@ -98,35 +98,6 @@ __license__ = "mit"
 
 
 LOGGER = logging.getLogger(__name__)
-
-
-def setup_project_folder(args):
-    """Creates folder structure in the target directory
-
-    Parameters
-    ----------
-    args
-    """
-    _recursive_overwrite('smif', 'sample_project', args.directory)
-    if args.directory == ".":
-        dirname = "the current directory"
-    else:
-        dirname = args.directory
-    LOGGER.info("Created sample project in %s", dirname)
-
-
-def _recursive_overwrite(pkg, src, dest):
-    if pkg_resources.resource_isdir(pkg, src):
-        if not os.path.isdir(dest):
-            os.makedirs(dest)
-        contents = pkg_resources.resource_listdir(pkg, src)
-        for item in contents:
-            _recursive_overwrite(pkg,
-                                 os.path.join(src, item),
-                                 os.path.join(dest, item))
-    else:
-        filename = pkg_resources.resource_filename(pkg, src)
-        shutil.copyfile(filename, dest)
 
 
 def list_model_runs(args):
